+++ backend/services/payment-service/repository/repository.go
@@ -8,7 +8,7 @@ import (
 
 type PaymentRepository interface {
 	FindAll(ctx context.Context, tx *sql.Tx) ([]*model.Payment, error)
-	FindAllByUserId(ctx context.Context, tx *sql.Tx, userId int) ([]*model.Payment, error)
+	FindAllByUserId(ctx context.Context, tx *sql.Tx, userId int64) ([]*model.Payment, error)
 	FindByOrderId(ctx context.Context, tx *sql.Tx, orderId string) (*model.Payment, error)
 	Create(ctx context.Context, tx *sql.Tx, payment *model.Payment) (*model.Payment, error)
 	Update(ctx context.Context, tx *sql.Tx, payment *model.Payment) error
